+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 This changelog goes through all the changes that have been made in each release. Information about changelogs can
 be found at [](https://keepachangelog.com).
 
+## [1.0.0-SNAPSHOT]() - 2023-04-30
+
+### Added
+* Add another stream api example
+
+### Changed
+* Changed User object builder
+
 ## [1.0.0-SNAPSHOT]() - 2023-03-01
 
 * Initial release